+++ tests/testthat/test-checkArgumentsFragPipe.R
@@ -149,8 +149,8 @@ test_that("argument checking for FP works", {
     ## species
     args <- args0
     args$species <- 1
-    expect_error(do.call(.checkArgumentsFragPipe, args),
-                 "Unknown species 1")
+    expect_warning(do.call(.checkArgumentsFragPipe, args),
+                   "Unknown species 1")
     args$species <- c("Mouse", "Human")
     expect_error(do.call(.checkArgumentsFragPipe, args),
                  "length(species) == 1 is not TRUE", fixed = TRUE)