+++ tests/testthat/test-checkArgumentsPDTMTptm.R
@@ -124,8 +124,8 @@ test_that("argument checking for PD-TMT PTM works", {
     ## species
     args <- args0
     args$species <- 1
-    expect_error(do.call(.checkArgumentsPDTMTptm, args),
-                 "Unknown species 1")
+    expect_warning(do.call(.checkArgumentsPDTMTptm, args),
+                   "Unknown species 1")
     args$species <- c("Mouse", "Human")
     expect_error(do.call(.checkArgumentsPDTMTptm, args),
                  "length(species) == 1 is not TRUE", fixed = TRUE)