+++ NEWS.md
@@ -6,6 +6,7 @@
 * Add argument to define the assays(s) to use for exported values and barplots
 * Harmonize treatment of merged groups in runTest and plotVolcano
 * Sort exported test results by p-value instead of logFC
+* Expand the set of allowed species
 
 # einprot 0.7.3